6-1. styling, 아직 진행 중(HomeStyle.js 부터)

+++ src/routes/Profile.js
@@ -36,6 +36,7 @@ const Profile = ({ refreshUser, userObj }) => {
       where("creatorId", "==", userObj.uid),
       orderBy("createdAt"),
     );
+    // eslint-disable-next-line no-unused-vars
     const querySnapshot = await getDocs(q);
     // querySnapshot.forEach((doc) => {
     //   console.log(doc.id, "=>", doc.data());
@@ -74,7 +75,7 @@ const Profile = ({ refreshUser, userObj }) => {
       refreshUser();
     }
     */
-    
+
     if (
       userObj.displayName !== newDisplayName ||
       userObj.photoURL !== attachment
@@ -84,12 +85,19 @@ const Profile = ({ refreshUser, userObj }) => {
       );
       if (ok) {
         // 기존 profile image 지우기, 현재 image url(userObj.photoURL)을 storage 안에서 자동으로 찾아 지우는 듯
-        const r = ref(storageService(), userObj.photoURL);
-        await deleteObject(r);
 
         try {
+          // 현재 내 프로필의 유무와 선택한 이미지파일 유무에 따라 storage에 저장할지 여부를 결정
           let attachmentUrl = "";
           if (attachment !== "") {
+            if (authService().currentUser.photoURL) {
+              const r = ref(
+                storageService(),
+                authService().currentUser.photoURL,
+              );
+              await deleteObject(r);
+            }
+
             const attachmentRef = ref(
               storageService(),
               `${userObj.uid}/profileImages/${uuidv4()}`,
@@ -101,19 +109,36 @@ const Profile = ({ refreshUser, userObj }) => {
               "data_url",
             );
             attachmentUrl = await getDownloadURL(attachmentRef);
+          } else {
+            const sure = window.confirm(
+              "Are you sure you want to update empty profile image?",
+            );
+            if (sure) {
+              if (authService().currentUser.photoURL) {
+                const r = ref(
+                  storageService(),
+                  authService().currentUser.photoURL,
+                );
+                await deleteObject(r);
+              }
+            } else {
+              alert("please again profile update!");
+              return;
+            }
           }
 
-          /* 5-2. 댓글 참고 수정 전 주석처리
-          await updateProfile(userObj, {
-            displayName: newDisplayName,
-            photoURL: attachmentUrl,
-          });
-          */
+          // 5-2. 댓글 참고 수정 전 주석처리
+          // 지금 상황에 이걸로 하면 getitorken?이 필요하다고 하는데 현재 userObj에는 App.js에서 내가 필요한 정보들만(displayName, photoURL, email, .. 등등) 있기 때문에 getitorken 뭐시기를 못찾아서 문제가 생긴다. 그래서 아래코드(authService().currentUser)로 사용
+          // await updateProfile(userObj, {
+          //   displayName: newDisplayName,
+          //   photoURL: attachmentUrl,
+          // });
+          // auth 정보 수정
           await updateProfile(authService().currentUser, {
             displayName: newDisplayName,
             photoURL: attachmentUrl,
           });
-          refreshUser();
+          // refreshUser();
           // 이렇게 했을 때 App.js의 setUserObj에 updateProfile을 새로 선언해주지 않아도 정상동작한다고 한다.
 
           const { email, uid } = userObj;
@@ -124,12 +149,8 @@ const Profile = ({ refreshUser, userObj }) => {
             uid: uid,
           };
           const newUser = doc(dbService(), "users", userObj.uid);
-          const docRef = await setDoc(
-            newUser,
-            newUserObj,
-            { capital: false },
-            { merge: false },
-          );
+          // eslint-disable-next-line no-unused-vars
+          const docRef = await setDoc(newUser, newUserObj);
 
           setAttachment(""); // 미리보기 사진 제거
           fileInput.current.value = ""; // 올리고 선택된 파일 해제
